4-21

+++ controllers/blogs.js
@@ -19,7 +19,6 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   try {
 
-    
     if (!request.token) {
       return response.status(401).json({ error: 'token missing' })
     }
@@ -51,6 +50,20 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   try {
+
+    if (!request.token) {
+      return response.status(401).json({ error: 'token missing' })
+    }
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'invalid token' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+      return response.status(403).json({ error: 'only blog owner can remove blogs'})
+    }
+    
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
   } catch (exception) {
@@ -61,6 +74,7 @@ blogsRouter.delete('/:id', async (request, response) => {
 
 blogsRouter.put('/:id', async (request, response) => {
   try {
+    
     const body = request.body
     const updatedBlog = await Blog
       .findByIdAndUpdate(request.params.id, {...body}, { new: true })
